refactor: move eventHandler.js file position to src/handlers/ folder
